# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dhasvin Enterprises Inventory Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: getSuppliers
      tags: [suppliers]
      summary: Get all suppliers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierRequest"
      responses:
        "201":
          description: Supplier created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get a supplier by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierWithProducts"
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierRequest"
      responses:
        "200":
          description: Supplier updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /sales:
    get:
      operationId: getSales
      tags: [sales]
      summary: Get all sales records
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sales records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Record a sale
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleRequest"
      responses:
        "201":
          description: Sale recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /purchases:
    get:
      operationId: getPurchases
      tags: [purchases]
      summary: Get all purchase records
      security:
        - bearerAuth: []
      parameters:
        - name: supplierId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Purchase records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Record a purchase
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseRequest"
      responses:
        "201":
          description: Purchase recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, staff, user]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, staff, user]
      required:
        - email
        - name
        - password
        - role
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, staff, user]
        password:
          type: string
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        gst:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
    SupplierWithProducts:
      allOf:
        - $ref: "#/components/schemas/Supplier"
        - type: object
          properties:
            products:
              type: array
              items:
                $ref: "#/components/schemas/Product"
            purchaseHistory:
              type: array
              items:
                $ref: "#/components/schemas/Purchase"
    CreateSupplierRequest:
      type: object
      properties:
        name:
          type: string
        gst:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
      required:
        - name
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        price:
          type: number
          format: float
        quantity:
          type: integer
        minQuantity:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        hsnCode:
          type: string
        gst:
          type: number
          format: float
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - price
        - quantity
...
    Sale:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: float
        totalAmount:
          type: number
          format: float
        discount:
          type: number
          format: float
        gst:
          type: number
          format: float
        customerName:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - quantity
        - unitPrice
        - totalAmount
    CreateSaleRequest:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
          format: float
        discount:
          type: number
          format: float
        gst:
          type: number
          format: float
        customerName:
          type: string
        notes:
          type: string
      required:
        - productId
        - quantity
        - unitPrice
    Purchase:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: float
        totalAmount:
          type: number
          format: float
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - quantity
        - unitPrice
    CreatePurchaseRequest:
      type: object
      properties:
        productId:
          type: integer
        supplierId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
          format: float
        notes:
          type: string
      required:
        - productId
        - quantity
        - unitPrice
...